add tableview to saved view
